+++ wallets/react-wallet-chat/src/pages/newChat.tsx
@@ -1,22 +1,27 @@
 import { Fragment, useCallback, useEffect, useState } from 'react'
 import { useSnapshot } from 'valtio'
 import { FiArrowRight } from 'react-icons/fi'
-import { HiQrcode } from 'react-icons/hi'
 import { Input, Row } from '@nextui-org/react'
-import { Web3Modal } from '@web3modal/standalone'
 import PageHeader from '@/components/PageHeader'
 import { chatClient } from '@/utils/WalletConnectUtil'
 import { ChatClientTypes } from '@walletconnect/chat-client'
 import ChatPrimaryCTAButton from '@/components/ChatPrimaryCTAButton'
 import { demoContactsMap } from '@/config/chatConstants'
 import SettingsStore from '@/store/SettingsStore'
-
-const web3modal = new Web3Modal({})
+import { useRouter } from 'next/router'
 
 export default function NewChatPage() {
   const [address, setAddress] = useState('')
+  const router = useRouter()
 
   const { eip155Address } = useSnapshot(SettingsStore.state)
+
+  useEffect(() => {
+    chatClient.on('chat_joined', args => {
+      const newChatTarget = new URLSearchParams(document.location.search).get('target')
+      router.push(`/chat?topic=${args.topic}&peerAccount=${newChatTarget}`)
+    })
+  }, [router])
 
   const createInvite = useCallback(
     async (targetAddress: string) => {
@@ -45,11 +50,6 @@ export default function NewChatPage() {
     [setAddress, createInvite]
   )
 
-  const inviteQrCode = useCallback(async () => {
-    const uri = `${window.location.origin}/newChat?accountId=1&target=eip155:1:${SettingsStore.state.eip155Address}`
-    web3modal.openModal({ uri })
-  }, [])
-
   useEffect(() => {
     const newChatTarget = new URLSearchParams(document.location.search).get('target')
     if (newChatTarget) {
@@ -65,10 +65,7 @@ export default function NewChatPage() {
         withBackButton
         backButtonHref="/chats"
         ctaButton={
-          <Row justify="space-evenly">
-            <ChatPrimaryCTAButton icon={<HiQrcode />} onClick={inviteQrCode} />
-            <ChatPrimaryCTAButton icon={<FiArrowRight />} onClick={() => onInvite(address)} />
-          </Row>
+          <ChatPrimaryCTAButton icon={<FiArrowRight />} onClick={() => onInvite(address)} />
         }
       />
